% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{data}
\name{acs}
\alias{acs}
\title{Demographic data of 857 patients with ACS}
\format{
An object of class \code{data.frame} with 857 rows and 17 columns.
}
\usage{
acs
}
\description{
A dataset containing demographic data and laboratory data of 857 patients with
acute coronary syndrome(ACS).
}
\examples{
interpretCI::acs
}
\keyword{datasets}
