% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_n.R
\name{draw_n}
\alias{draw_n}
\title{Draw normal distribution curve}
\usage{
draw_n(mean = 0, sd = 1, z = NULL, p = 0.05, alternative = "two.sided")
}
\arguments{
\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{z}{vector of quantiles}

\item{p}{vector of probabilities}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}
}
\value{
A ggplot
}
\description{
Draw normal distribution curve
}
\examples{
draw_n()
draw_n(alternative="less")
draw_n(alternative="greater")
draw_n(z=-1.75)
draw_n(z=-1.75,alternative="greater")
draw_n(z=-1.75,alternative="less")
}
