% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_t.R
\name{draw_t}
\alias{draw_t}
\title{Draw t distribution curve}
\usage{
draw_t(DF = 50, t = NULL, p = 0.05, alternative = "two.sided")
}
\arguments{
\item{DF}{numeric degree of freedom}

\item{t}{numeric t value}

\item{p}{numeric p value}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}
}
\value{
A ggplot
}
\description{
Draw t distribution curve
}
\examples{
draw_t(DF=30)
draw_t(DF=20,t=2.5)
draw_t(DF=49,t=1.77)
draw_t(DF=49,p=0.005)
draw_t(DF=19,t=-0.894,alternative="less")
draw_t(DF=146,t=0.67,alternative="greater")
}
