% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_x2.R
\name{draw_x2}
\alias{draw_x2}
\title{Draw chi-squared distribution curve}
\usage{
draw_x2(q = NULL, p = 0.05, df = 2)
}
\arguments{
\item{q}{vector of quantiles}

\item{p}{vector of probabilities}

\item{df}{degrees of freedom (non-negative, but can be non-integer).}
}
\value{
A ggplot
}
\description{
Draw chi-squared distribution curve
}
\examples{
draw_x2(df=1)
draw_x2(q=0.78)
}
