% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english2.R
\name{English}
\alias{English}
\title{Convert numeric to string with uppercase first letter}
\usage{
English(x, digits = 2)
}
\arguments{
\item{x}{A numeric}

\item{digits}{integer indicating the number of decimal places}
}
\value{
A string
}
\description{
Convert numeric to string with uppercase first letter
}
\examples{
English(40)
English(13.1)
}
