% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/english2.R
\name{english2}
\alias{english2}
\title{Convert numeric to string}
\usage{
english2(x, digits = 2)
}
\arguments{
\item{x}{A numeric}

\item{digits}{integer indicating the number of decimal places}
}
\value{
A character string
}
\description{
Convert numeric to string
}
\examples{
english2(45)
english2(12.34)
}
