% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meanCI.R
\name{estimationPlot1}
\alias{estimationPlot1}
\title{Draw estimation plot1}
\usage{
estimationPlot1(x, palette = NULL)
}
\arguments{
\item{x}{An object of class meanCI}

\item{palette}{The name of color palette from RColorBrewer package or NULL}
}
\value{
A ggplot
}
\description{
Draw estimation plot1
}
\examples{
x=meanCI(iris,Species,Sepal.Length)
estimationPlot1(x)
}
