% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interprete.R
\name{interprete}
\alias{interprete}
\title{Interprete an object of meanCI}
\usage{
interprete(x, viewer = "rstudio")
}
\arguments{
\item{x}{An object of class "meanCI"}

\item{viewer}{Character One of c("rstudio","browser")}
}
\description{
Interprete an object of meanCI
}
\examples{
x=meanCI(mtcars$mpg)
x=meanCI(mtcars,mpg,mu=23)
x=meanCI(n=150,m=115,s=10,alpha=0.01)
x=meanCI(n=50,m=295,s=20,mu=300)
x= meanCI(n=20,m=108,s=10,mu=110,alpha=0.01,alternative="less")
\dontrun{
x=meanCI(n1=500,n2=1000,m1=20,s1=3,m2=15,s2=2,alpha=0.01)
x=meanCI(n1=15,n2=20,m1=1000,s1=100,m2=950,s2=90,alpha=0.1)
x=meanCI(n1=30,n2=25,m1=78,s1=10,m2=85,s2=15,mu=0,alpha=0.10)
x=meanCI(n1=100,n2=100,m1=200,s1=40,m2=190,s2=20,mu=7,alpha=0.05,alternative="greater")
x1=c(95,89,76,92,91,53,67,88,75,85,90,85,87,85,85,68,81,84,71,46,75,80)
y1=c(90,85,73,90,90,53,68,90,78,89,95,83,83,83,82,65,79,83,60,47,77,83)
x=meanCI(x=x1,y=y1,paired=TRUE,alpha=0.1,mu=0)
x=propCI(n=1600,p=0.4,alpha=0.01)
x=propCI(n=100,p=0.73,P=0.8,alpha=0.01)
x=propCI(n=100,p=0.73,P=0.8,alpha=0.05,alternative="greater")
x=propCI(n1=100,n2=200,p1=0.38,p2=0.51,alpha=0.01)
x=propCI(n1=150,n2=100,p1=0.71,p2=0.63,P=0,alternative="greater")
interprete(x)
interprete(x,"browser")
}
}
