% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{isProvided}
\alias{isProvided}
\title{Whether the arg is provided in function call}
\usage{
isProvided(x, seek = "mu")
}
\arguments{
\item{x}{An object of class "meanCI" or function call or character string}

\item{seek}{character. Default="mu"}
}
\value{
logical
}
\description{
Whether the arg is provided in function call
}
\examples{
x=meanCI(mtcars,am,mpg)
isProvided(x)
}
