% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCI2df.R
\name{meanCI2df}
\alias{meanCI2df}
\title{Prepare data to plot from an object of class meanCI}
\usage{
meanCI2df(x)
}
\arguments{
\item{x}{An object of class meanCI}
}
\value{
A data.frame
}
\description{
Prepare data to plot from an object of class meanCI
}
\examples{
x=meanCI(acs,sex,age)
meanCI2df(x)
}
