% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCI2df1.R
\name{meanCI2df1}
\alias{meanCI2df1}
\title{Extract data from an object of class meanCI}
\usage{
meanCI2df1(x)
}
\arguments{
\item{x}{An object of class meanCI}
}
\value{
A data.frame summarizing mean and confidence interval
}
\description{
Extract data from an object of class meanCI
}
\examples{
x=meanCI(acs,sex,age)
meanCI2df1(x)
}
