% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meanCI.R
\name{meanCIplot1}
\alias{meanCIplot1}
\title{Draw meanCI plot for data with single vector}
\usage{
meanCIplot1(x)
}
\arguments{
\item{x}{An object of class "meanCI" with attr(x,"measure")=="mean"}
}
\value{
A ggplot
}
\description{
Draw meanCI plot for data with single vector
}
\examples{
x=meanCI(mtcars,mpg)
meanCIplot1(x)
}
