% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meanCI.R
\name{pairPlot}
\alias{pairPlot}
\title{Draw a pair plot with an object of class meanCI}
\usage{
pairPlot(x, palette = NULL)
}
\arguments{
\item{x}{An object of class "meanCI" with attr(x,"measure")=="paired"}

\item{palette}{The name of color palette from RColorBrewer package or NULL}
}
\value{
A ggplot
}
\description{
Draw a pair plot with an object of class meanCI
}
\examples{
x=meanCI(iris,paired=TRUE)
pairPlot(x)
x=meanCI(iris,Petal.Width, Petal.Length,paired=TRUE)
pairPlot(x)
}
