% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meanCI.R
\name{pairPlot1}
\alias{pairPlot1}
\title{Draw a pair plot}
\usage{
pairPlot1(data, ref = NULL, palette = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{ref}{Numeric or NULL}

\item{palette}{The name of color palette from RColorBrewer package or NULL}
}
\value{
A ggplot
}
\description{
Draw a pair plot
}
\examples{
x=meanCI(mtcars,paired=TRUE)
pairPlot1(x$data)
pairPlot1(x$data,ref=c(1,4,6))
pairPlot1(x$data,ref=c(1,3))
}
