% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette2colors.R
\name{palette2colors}
\alias{palette2colors}
\title{Extract hexadecimal colors from a color palette}
\usage{
palette2colors(name, reverse = FALSE)
}
\arguments{
\item{name}{The name of color palette from RColorBrewer package}

\item{reverse}{Whether or not reverse the order of colors}
}
\value{
hexadecimal colors
}
\description{
Extract hexadecimal colors from a color palette
}
\examples{
palette2colors("Reds")
}
