% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.meanCI.R
\name{print.meanCI}
\alias{print.meanCI}
\title{S3 method "print" for class "meanCI"}
\usage{
\method{print}{meanCI}(x, ...)
}
\arguments{
\item{x}{An object of class "meanCI"}

\item{...}{Further arguments}
}
\description{
S3 method "print" for class "meanCI"
}
