% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meanCI.R
\name{print.plotCI}
\alias{print.plotCI}
\title{S3 method for class plotCI}
\usage{
\method{print}{plotCI}(x, ...)
}
\arguments{
\item{x}{An object of class plotCI}

\item{...}{Further arguments}
}
\value{
No return value, called for side effect
}
\description{
S3 method for class plotCI
}
