% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/English.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{english}
\alias{\%>\%}
\alias{select}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr]{select}}}

  \item{english}{\code{\link[english:as.english]{english}}}
}}

