% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interprete.R
\name{rstudio_viewer}
\alias{rstudio_viewer}
\title{Show html file in RStudio viewer or browser}
\usage{
rstudio_viewer(file_name, file_path = NULL, viewer = "rstudio")
}
\arguments{
\item{file_name}{character file name}

\item{file_path}{character file path}

\item{viewer}{Character One of c("rstudio","browser")}
}
\description{
Show html file in RStudio viewer or browser
}
