% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_t_table.R
\name{show_t_table}
\alias{show_t_table}
\title{Show t-value table}
\usage{
show_t_table(DF = 20, t = NULL, p = 0.05, alternative = "two.sided")
}
\arguments{
\item{DF}{Numeric degree of freedom}

\item{t}{Numeric vector of quantile}

\item{p}{Numeric probability}

\item{alternative}{Character One of c("two.sided","greater","less")}
}
\value{
An object of class "flextable"
}
\description{
Show t-value table
}
\examples{
show_t_table()
show_t_table(t=1.4)
show_t_table(DF=10)
}
