% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_t_table.R
\name{show_x2_table}
\alias{show_x2_table}
\title{Show chisquare table}
\usage{
show_x2_table(DF = 1, x2 = NULL, p = 0.05)
}
\arguments{
\item{DF}{Numeric degree of freedom}

\item{x2}{Numeric vector of chi-square value}

\item{p}{Numeric probability}
}
\value{
An object of class "flextable"
}
\description{
Show chisquare table
}
\examples{
show_x2_table(DF=2,x2=1.5)
show_x2_table(p=0.05)
}
