% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textBox.R
\name{textBox}
\alias{textBox}
\title{Draw textbox}
\usage{
textBox(
  string,
  color = "black",
  lcolor = "red",
  bg = "cornsilk",
  lwd = 1,
  width = 10,
  bold = FALSE,
  italic = FALSE,
  fontsize = 11,
  space = 1.5,
  fontname
)
}
\arguments{
\item{string}{string}

\item{color}{font color}

\item{lcolor}{line color}

\item{bg}{background color}

\item{lwd}{numeric line width}

\item{width}{numeric box width}

\item{bold, italic}{logical}

\item{fontsize}{numeric font size}

\item{space}{space between lines}

\item{fontname}{name of font}
}
\value{
A flextable
}
\description{
Draw textbox
}
\examples{
string="Good Morning!"
textBox(string,italic=TRUE)
}
