% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/English.R
\name{English}
\alias{English}
\title{Convert numeric to string with uppercase first letter}
\usage{
English(x)
}
\arguments{
\item{x}{A numeric}
}
\value{
A string
}
\description{
Convert numeric to string with uppercase first letter
}
