% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.meanCI.R
\name{pairPlot}
\alias{pairPlot}
\title{Draw a pair plot}
\usage{
pairPlot(data, ref = NULL, palette = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{ref}{Numeric or NULL}

\item{palette}{The name of color palette from RColorBrewer package or NULL}
}
\value{
a ggplot
}
\description{
Draw a pair plot
}
\examples{
x=meanCI(mtcars,paired=TRUE)
pairPlot(x$data)
pairPlot(x$data,ref=c(1,4,6))
pairPlot(x$data,ref=c(1,3))
}
