% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propCI.R
\name{propCI_sub}
\alias{propCI_sub}
\title{Calculate confidence intervals of proportion or difference between proportions in a data.frame}
\usage{
propCI_sub(data, x, y = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{x}{Character Name of a categorical column}

\item{y}{Character Optional. Name of another categorical column}
}
\description{
Calculate confidence intervals of proportion or difference between proportions in a data.frame
}
\examples{
propCI_sub(acs,"sex")
propCI_sub(acs,"sex","HBP")
}
